+++ src/test/java/domain/BlackJackGameTest.java
@@ -6,6 +6,7 @@ import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardShape;
 import domain.card.Hand;
+import domain.card.StaticCardGenerator;
 import domain.participant.Dealer;
 import domain.participant.Money;
 import domain.participant.Player;
