+++ src/test/java/domain/participant/PlayerTest.java
@@ -1,11 +1,13 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.GameResult;
 import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardShape;
 import domain.card.Hand;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
@@ -14,67 +16,20 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class HandTest {
-
-    @DisplayName("카드를 추가할 수 있다")
+public class PlayerTest {
+    @DisplayName("카드를 추가할 수 있다.")
     @Test
-    void test1() {
-        //given
-        Hand hand = Hand.empty();
-        Card card = new Card(CardNumber.TWO, CardShape.CLOVER);
+    void 카드_추가_가능() {
+        // given
+        Player player = Player.init("플레이어", Money.of("100000"));
+        Card card = new Card(CardNumber.A, CardShape.CLOVER);
+        player.addCard(card);
 
-        Hand expected = Hand.of(List.of(card));
-        //when
-        hand.add(card);
-        //then
-        assertThat(hand).isEqualTo(expected);
-    }
+        // when
+        Hand hand = player.getHand();
 
-    @DisplayName("숫자 카드의 합을 구할 수 있다")
-    @ParameterizedTest
-    @MethodSource("createNumberCardsCase")
-    void 숫자_카드_합_구하기(CardNumber number1, CardNumber number2, int expected) {
-        //given
-        Card card1 = new Card(number1, CardShape.CLOVER);
-        Card card2 = new Card(number2, CardShape.CLOVER);
-        Hand hand = Hand.of(List.of(card1, card2));
-        //when
-        int actual = hand.calculateFinalScore();
-        //then
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    private static Stream createNumberCardsCase() {
-        return Stream.of(
-                Arguments.of(CardNumber.TWO, CardNumber.FOUR, 6),
-                Arguments.of(CardNumber.TWO, CardNumber.THREE, 5),
-                Arguments.of(CardNumber.TWO, CardNumber.TWO, 4),
-                Arguments.of(CardNumber.TWO, CardNumber.A, 13)
-        );
-    }
-
-    @DisplayName("A를 제외한 문자 카드의 합을 구할 수 있다")
-    @ParameterizedTest
-    @MethodSource("createCharCardsCase")
-    void 문자_카드_합_구하기(CardNumber number1, CardNumber number2, int expected) {
-        //given
-        Card card1 = new Card(number1, CardShape.CLOVER);
-        Card card2 = new Card(number2, CardShape.CLOVER);
-        Hand hand = Hand.of(List.of(card1, card2));
-
-        //when
-        int actual = hand.calculateFinalScore();
-
-        //then
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    private static Stream createCharCardsCase() {
-        return Stream.of(
-                Arguments.of(CardNumber.KING, CardNumber.THREE, 13),
-                Arguments.of(CardNumber.JACK, CardNumber.TEN, 20),
-                Arguments.of(CardNumber.QUEEN, CardNumber.KING, 20)
-        );
+        // then
+        assertThat(hand.getCards()).contains(card);
     }
 
     @DisplayName("21 이하일 때, 최적의 결과를 선택할 수 있다.")
@@ -83,12 +38,13 @@ public class HandTest {
     void 최적_결과_선택_21_이하(List<Card> inputCards, int expected) {
         // given
         Hand hand = Hand.of(inputCards);
+        Player player = Player.of(hand, "플레이어1", Money.of("100000"));
 
         // when
-        final int score = hand.calculateFinalScore();
+        final int result = player.calculateScore();
 
         // then
-        assertThat(score).isEqualTo(expected);
+        assertThat(result).isEqualTo(expected);
     }
 
     private static Stream createCardsCase() {
@@ -122,14 +78,15 @@ public class HandTest {
     @ParameterizedTest
     @MethodSource("createBurstCardsCase")
     void 가장_가까운_값_선택(List<Card> inputCards, int expected) {
-        // given
+        //given
         Hand hand = Hand.of(inputCards);
+        Player player = Player.of(hand, "플레이어1", Money.of("100000"));
 
-        // when
-        final int score = hand.calculateFinalScore();
+        //when
+        int actual = player.calculateScore();
 
         //then
-        assertThat(score).isEqualTo(expected);
+        assertThat(actual).isEqualTo(expected);
     }
 
     private static Stream createBurstCardsCase() {
@@ -155,16 +112,15 @@ public class HandTest {
         );
     }
 
-    @DisplayName("플레이어가 소유한 카드에 따라서 bust 여부를 판단한다.")
+    @DisplayName("플레이어가 소유한 카드에 따라서 burst 여부를 판단한다.")
     @ParameterizedTest
     @MethodSource("createBurstCase")
     void test1(List<Card> inputCard, boolean expected) {
-        // given
+        //given
         Hand hand = Hand.of(inputCard);
-
-        // when
-        final boolean actual = hand.isBust();
-
+        Player dealer = Player.of(hand, "플레이어1", Money.of("100000"));
+        //when
+        final boolean actual = dealer.isBurst();
         //then
         assertThat(actual).isEqualTo(expected);
     }
@@ -190,71 +146,80 @@ public class HandTest {
         );
     }
 
-    @DisplayName("플레이어가 소유한 카드에 따라서 21 여부를 판단한다.")
-    @ParameterizedTest
-    @MethodSource("create21Case")
-    void 합이_21인지_검증(List<Card> inputCard, boolean expected) {
-        // given
-        Hand hand = Hand.of(inputCard);
-
-        // when
-        final boolean actual = hand.isTwentyOne();
-
-        //then
-        assertThat(actual).isEqualTo(expected);
-    }
-
-    private static Stream<Arguments> create21Case() {
-        return Stream.of(
-                Arguments.of(
-                        List.of(
-                                new Card(CardNumber.A, CardShape.CLOVER),
-                                new Card(CardNumber.TEN, CardShape.CLOVER)
-                        ),
-                        true
-                ),
-                Arguments.of(
-                        List.of(
-                                new Card(CardNumber.A, CardShape.CLOVER),
-                                new Card(CardNumber.TEN, CardShape.CLOVER),
-                                new Card(CardNumber.A, CardShape.CLOVER)
-                        ),
-                        false
-                )
-        );
-    }
-
-    @DisplayName("플레이어가 소유한 카드에 따라서 블랙잭 여부를 판단한다.")
+    @DisplayName("플레이어가 가진 카드의 블랙잭 여부를 판단한다.")
     @ParameterizedTest
     @MethodSource("createBlackJackCase")
-    void 블랙잭인지_검증(List<Card> inputCard, boolean expected) {
-        // given
+    void test2(List<Card> inputCard, boolean expected) {
+        //given
         Hand hand = Hand.of(inputCard);
-
-        // when
-        final boolean actual = hand.isBlackJack();
-
+        Player dealer = Player.of(hand, "플레이어1", Money.of("100000"));
+        //when
+        final boolean actual = dealer.isBlackJack();
         //then
         assertThat(actual).isEqualTo(expected);
     }
 
-    private static Stream<Arguments> createBlackJackCase() {
+    private static Stream createBlackJackCase() {
         return Stream.of(
                 Arguments.of(
                         List.of(
-                                new Card(CardNumber.A, CardShape.CLOVER),
-                                new Card(CardNumber.TEN, CardShape.CLOVER)
+                                new Card(CardNumber.TEN, CardShape.CLOVER),
+                                new Card(CardNumber.A, CardShape.CLOVER)
                         ),
                         true
                 ),
                 Arguments.of(
                         List.of(
                                 new Card(CardNumber.A, CardShape.CLOVER),
-                                new Card(CardNumber.TEN, CardShape.CLOVER),
-                                new Card(CardNumber.A, CardShape.CLOVER)
+                                new Card(CardNumber.QUEEN, CardShape.CLOVER),
+                                new Card(CardNumber.TWO, CardShape.CLOVER)
                         ),
                         false
                 )
         );
+    }
+
+    @Test
+    void 플레이어가_이겼을_경우_수익을_계산한다() {
+        //given
+        Hand hand = Hand.of(new ArrayList<>());
+        Player player = Player.of(hand, "플레이어1", Money.of("100000"));
+        //when
+        Money actual = player.calculateRevenue(GameResult.WIN);
+        //then
+        assertThat(actual).isEqualTo(Money.of(100000));
+    }
+
+    @Test
+    void 플레이어가_블랙잭으로_이겼을_경우_수익을_계산한다() {
+        //given
+        Hand hand = Hand.of(new ArrayList<>());
+        Player player = Player.of(hand, "플레이어1", Money.of("100000"));
+        //when
+        Money actual = player.calculateRevenue(GameResult.BLACKJACK);
+        //then
+        assertThat(actual).isEqualTo(Money.of(150000));
+    }
+
+    @Test
+    void 플레이어가_비겼을_경우_수익을_계산한다() {
+        //given
+        Hand hand = Hand.of(new ArrayList<>());
+        Player player = Player.of(hand, "플레이어1", Money.of("100000"));
+        //when
+        Money actual = player.calculateRevenue(GameResult.DRAW);
+        //then
+        assertThat(actual).isEqualTo(Money.of(0));
+    }
+
+    @Test
+    void 플레이어가_졌을_경우_수익을_계산한다() {
+        //given
+        Hand hand = Hand.of(new ArrayList<>());
+        Player player = Player.of(hand, "플레이어1", Money.of("100000"));
+        //when
+        Money actual = player.calculateRevenue(GameResult.LOSE);
+        //then
+        assertThat(actual).isEqualTo(Money.of(-100000));
     }
 }
