+++ src/test/java/domain/participant/PlayersTest.java
@@ -1,10 +1,7 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.participant.Money;
-import domain.participant.Player;
-import domain.participant.Players;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
